requirements.txt
fastapi
uvicorn
python-dotenv
spacy
beautifulsoup4
emoji
transformers
torch
